+++ src/components/DataTable/index.tsx
@@ -14,7 +14,7 @@ import { EmptyState, EmptyStateProps } from '~/components/EmptyState';
 import { SortObject } from '~/types';
 import TableCellBody from '~/components/TableCellBody';
 import TableCellHead from '~/components/TableCellHead';
-import TableContainerPrimary from '~/components/TableContainerPrimary';
+import TableContainer from '~/components/TableContainer';
 
 import useStyles from './styles';
 
@@ -37,7 +37,7 @@ export interface DataTableProps<T> {
   ) => React.ReactElement<any>;
 }
 
-const DataTable = <T,>({
+export const DataTable = <T,>({
   columns,
   currentSortObject,
   emptyStateProps,
@@ -48,7 +48,7 @@ const DataTable = <T,>({
   const classes = useStyles();
 
   return (
-    <TableContainerPrimary>
+    <TableContainer>
       <Table>
         <TableHead>
           <TableRow>
@@ -69,32 +69,27 @@ const DataTable = <T,>({
         </TableHead>
 
         <TableBody>
-          {items.map(
-            (item, i): React.ReactElement<any> => {
-              /* eslint-disable react/no-array-index-key */
-              const children = (
-                <TableRow hover key={i}>
-                  {columns.map(
-                    (column, y): React.ReactElement<any> => (
-                      <TableCellBody
-                        key={column.label || y}
-                        style={column.style}
-                      >
-                        {column.renderValue(item)}
-                      </TableCellBody>
-                    ),
-                  )}
-                </TableRow>
-              );
-              /* eslint-enable react/no-array-index-key */
+          {items.map((item, i): React.ReactElement<any> => {
+            /* eslint-disable react/no-array-index-key */
+            const children = (
+              <TableRow hover key={i}>
+                {columns.map(
+                  (column, y): React.ReactElement<any> => (
+                    <TableCellBody key={column.label || y} style={column.style}>
+                      {column.renderValue(item)}
+                    </TableCellBody>
+                  ),
+                )}
+              </TableRow>
+            );
+            /* eslint-enable react/no-array-index-key */
 
-              if (RowWrapper) {
-                return RowWrapper(item, children);
-              }
+            if (RowWrapper) {
+              return RowWrapper(item, children);
+            }
 
-              return children;
-            },
-          )}
+            return children;
+          })}
         </TableBody>
       </Table>
 
@@ -103,7 +98,7 @@ const DataTable = <T,>({
           <EmptyState description="No Results Found" {...emptyStateProps} />
         </div>
       )}
-    </TableContainerPrimary>
+    </TableContainer>
   );
 };