+++ CHANGELOG.md
@@ -10,7 +10,9 @@ import directly from Material UI, such as `<Button />` and `<Link />`.
 importing a MUI component from the DLS vs. MUI. The only exception is with the
 `<Link />` component, as the MUI `Link` component does not accept a `to` prop.
 * Removed `Primary` from all component names:
-  * Renamed `<DataTablePrimary />` to `<DataTable >`
+  * Renamed `<DataTablePrimary />` to `<DataTable />`
+  * Renamed `<LoadingPrimary />` to `<Loading />`
+  * Renamed `<TableContainerPrimary />` to `<TableContainer />`
 
 ## [v4.1.0] - May 25, 2021